django_admin/requirements.txt
Django==5.2
djangorestframework==3.15.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
djangorestframework
djangorestframework-simplejwt

